+++ src/panels/SkyBrowserPanel/WorldWideTelescope/hooks.ts
@@ -1,5 +1,8 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
+import { useGesture } from '@use-gesture/react';
 
+import { useOpenSpaceApi } from '@/api/hooks';
+import { useBoolProperty } from '@/hooks/properties';
 import { useAppSelector } from '@/redux/hooks';
 import { useWindowSize } from '@/windowmanagement/Window/hooks';
 
@@ -11,7 +14,7 @@ import {
   useOpacities,
   useSelectedImages
 } from '../hooks';
-import { Status } from '../types';
+import { WwtStatus } from '../types';
 
 import { useWwtProvider } from './WwtProvider/hooks';
 
@@ -112,7 +115,7 @@ export function useUpdateBorderRadius(id: string) {
   }, [borderRadius, setBorderRadius, wwtHasLoaded, width, height]);
 }
 
-export function useOverlayStatus(): { visible: boolean; type: Status } {
+export function useOverlayStatus(): { visible: boolean; type: WwtStatus } {
   const cameraInSolarSystem = useAppSelector(
     (state) => state.skybrowser.cameraInSolarSystem
   );
@@ -129,4 +132,89 @@ export function useOverlayStatus(): { visible: boolean; type: Status } {
   } else {
     return { visible: false, type: 'Ok' };
   }
+}
+
+export function useWwtInteraction(id: string, width: number, height: number) {
+  const [isDragging, setIsDragging] = useState(false);
+  const [startDragPosition, setStartDragPosition] = useState({ x: 0, y: 0 });
+
+  const [inverseZoom] = useBoolProperty('Modules.SkyBrowser.InverseZoomDirection');
+  const luaApi = useOpenSpaceApi();
+
+  /**
+   * The `offset` parameter represents the accumulated distance of the drag gesture
+   * in pixels. It is an array where:
+   * - `offset[0]` corresponds to the horizontal distance (x-axis).
+   * - `offset[1]` corresponds to the vertical distance (y-axis).
+   *
+   * This value is updated as the user drags, providing the total displacement
+   * from the starting point of the gesture.
+   */
+  const bindGestures = useGesture(
+    {
+      onDrag: (state) => {
+        state.event.preventDefault();
+        if (!state.wheeling && !state.pinching && state.touches === 1) {
+          handleDrag(state.offset[0], state.offset[1]);
+        }
+      },
+      onDragStart: (state) => {
+        if (state.touches === 1 && !state.wheeling && !state.pinching) {
+          mouseDown(state.offset[0], state.offset[1]);
+        }
+      },
+      onDragEnd: () => mouseUp(),
+      onPinch: (state) => {
+        if (state.touches == 2) {
+          // The first direction determines the "scroll" direction
+          const [direction] = state.direction;
+          scroll(direction < 0 ? 50 : -50);
+        }
+      },
+      onWheel: (state) => {
+        scroll(state.event.deltaY);
+      }
+    },
+    {
+      // To make it differentiate better between dragging and pinching
+      drag: { threshold: 0.1 },
+      eventOptions: { passive: false }
+    }
+  );
+
+  function handleDrag(x: number, y: number) {
+    if (!id || !isDragging) {
+      return;
+    }
+    // Calculate pixel translation
+    const dx = x - startDragPosition.x;
+    const dy = y - startDragPosition.y;
+    // Call lua function with relative values
+    luaApi?.skybrowser.finetuneTargetPosition(id, [dx / width, dy / height]);
+  }
+
+  function mouseDown(x: number, y: number) {
+    if (!id) {
+      return;
+    }
+    luaApi?.skybrowser.startFinetuningTarget(id);
+    luaApi?.skybrowser.stopAnimations(id);
+    setIsDragging(true);
+    setStartDragPosition({ x, y });
+  }
+
+  function mouseUp() {
+    setIsDragging(false);
+  }
+
+  function scroll(deltaY: number) {
+    if (!id) {
+      return;
+    }
+    const scrollDirection = inverseZoom ? -deltaY : deltaY;
+    luaApi?.skybrowser.scrollOverBrowser(id, scrollDirection);
+    luaApi?.skybrowser.stopAnimations(id);
+  }
+
+  return { bindGestures, isDragging };
 }
